# pagarte-web/src/api/generated/models/ApiResponse.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type ApiResponse = {
    success?: boolean;
    message?: string | null;
};