# pagarte-web/src/api/generated/models/ApiResponseOfstring.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type ApiResponseOfstring = {
    data?: string | null;
    success?: boolean;
    message?: string | null;
};