# pagarte-web/src/api/generated/models/LoginRequest.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type LoginRequest = {
    usernameOrEmail?: string;
    password?: string;
};