# pagarte-web/src/api/generated/models/RegisterUserRequest.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type RegisterUserRequest = {
    username?: string;
    password?: string;
    email?: string;
};